リスト3.1-3.7

+++ src/ar/ardump.c
@@ -7,6 +7,7 @@
 #include <ar.h>
 #include <sys/stat.h>
 #include <sys/mman.h>
+#include "../linuxelf.h"
 
 static int elfdump(char *head)
 {
@@ -73,7 +74,7 @@ static int elfdump(char *head)
     for (j = 0; j < sym->sh_size / sym->sh_entsize; j++) {
       symp = (Elf_Sym *)(head + sym->sh_offset + sym->sh_entsize * j);
       if (!symp->st_name) continue;
-      printf("\t[%d]\t%d\t%d\t%s\n",
+      printf("\t[%d]\t%d\t%ld\t%s\n",
 	     j, (int)ELF_ST_TYPE(symp->st_info), symp->st_size,
 	     (char *)(head + str->sh_offset + symp->st_name));
     }
@@ -90,7 +91,7 @@ static int elfdump(char *head)
       symp = (Elf_Sym *)(head + sym->sh_offset +
 			 (sym->sh_entsize * ELF_R_SYM(relp->r_info)));
       if (!symp->st_name) continue;
-      printf("\t[%d]\t%d\t%s\n",
+      printf("\t[%d]\t%ld\t%s\n",
 	     j, ELF_R_SYM(relp->r_info),
 	     (char *)(head + str->sh_offset + symp->st_name));
     }
@@ -132,7 +133,7 @@ int main(int argc, char *argv[])
 	} else {
 	  fname = arhdr->ar_name;
 	}
-	fprintf(stderr, "found: %.*s\n", strchr(fname, '/') - fname, fname);
+	fprintf(stderr, "found: %.*s\n", (int)(strchr(fname, '/') - fname), fname);
 	elfdump((char *)arhdr + sizeof(struct ar_hdr));
       }
       if (size % 2) size++;
